Working implementation of log in
For the moment login works as before the react refactoring. Login modal and iframe stay in
index.html. Log in logic and css are in login-flow.js and login.css which are directly
imported in index.html. This should probably be refactored to react in the future.
This commit also includes some style editing to accomodate login links.