Refactor save and load backend sessions
Now backend load resource needs a user_id and a session_id parameter.
A new resource called 'available' has been added that returns a list of available sessions (incuding user_id parameter).
Restore session in ui now loads the last stored session and only works for anonymous user.
Also storing history of versions has been removed for now (to save space).
Next step is to show a list of available sessions to the user (returned by the available resource), then user clicks on
one of the sessions and we can construct the corresonding url.
Because when loading only user_id and session_id are needed but user does not need to be logged,
universal sharing of sessions will be possible.