Refactor delete sound from path
Delete sound from path now uses idx rather than soundID.
Using soundID was not correct as a path might have multiple instances of the same sound.
However, when sounds are removed (REMOVE_SOUND action) then these are removed by ID (removing all instances).
This fix does not solve issue #12, as the problem in this issue lies in the fact that PathListSound object's sound prop does not get updated when the sound is removed from the path (probably because the sound is not changed...).

https://github.com/ffont/freesound-explorer/issues/12